# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/login":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"tags":["auth"],"summary":"Register User","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["users"],"summary":"Read Users","operationId":"read_users_users__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"},"title":"Response Read Users Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{user_id}/promote-to-admin":{"post":{"tags":["users"],"summary":"Promote To Admin","operationId":"promote_to_admin_users__user_id__promote_to_admin_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/ban":{"post":{"tags":["users"],"summary":"Ban User","operationId":"ban_user_users__user_id__ban_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/unban":{"post":{"tags":["users"],"summary":"Unban User","operationId":"unban_user_users__user_id__unban_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/":{"post":{"tags":["ads"],"summary":"Create Ad","operationId":"create_ad_ads__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ad"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["ads"],"summary":"Get Ads","operationId":"get_ads_ads__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ad"},"title":"Response Get Ads Ads  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{ad_id}":{"get":{"tags":["ads"],"summary":"Get Ad","operationId":"get_ad_ads__ad_id__get","parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ad"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["ads"],"summary":"Delete Ad","operationId":"delete_ad_ads__ad_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{ad_id}/reviews":{"get":{"tags":["ads"],"summary":"Get Reviews By Ad Id","operationId":"get_reviews_by_ad_id_ads__ad_id__reviews_get","parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Get Reviews By Ad Id Ads  Ad Id  Reviews Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["ads"],"summary":"Create Review","operationId":"create_review_ads__ad_id__reviews_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{ad_id}/complaints":{"get":{"tags":["ads"],"summary":"Get Complaints By Ad Id","operationId":"get_complaints_by_ad_id_ads__ad_id__complaints_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Complaint"},"title":"Response Get Complaints By Ad Id Ads  Ad Id  Complaints Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["ads"],"summary":"Create Complaint","operationId":"create_complaint_ads__ad_id__complaints_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Complaint"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/reviews/{review_id}":{"get":{"tags":["ads"],"summary":"Get Review By Id","operationId":"get_review_by_id_ads_reviews__review_id__get","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["ads"],"summary":"Delete Review","operationId":"delete_review_ads_reviews__review_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Ad":{"properties":{"type":{"$ref":"#/components/schemas/AdType","default":"Продажа"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"price":{"type":"number","title":"Price"},"image_url":{"type":"string","title":"Image Url"},"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","description","price","image_url","id","user_id","created_at"],"title":"Ad"},"AdCreate":{"properties":{"type":{"$ref":"#/components/schemas/AdType","default":"Продажа"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"price":{"type":"number","title":"Price"},"image_url":{"type":"string","title":"Image Url"}},"type":"object","required":["name","description","price","image_url"],"title":"AdCreate"},"AdType":{"type":"string","enum":["Продажа","Покупка","Услуга"],"title":"AdType"},"Body_login_for_access_token_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_login_post"},"Body_login_for_access_token_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_token_post"},"Complaint":{"properties":{"text":{"type":"string","title":"Text"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"id":{"type":"integer","title":"Id"},"ad_id":{"type":"integer","title":"Ad Id"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["text","id","ad_id","user_id","created_at"],"title":"Complaint"},"ComplaintCreate":{"properties":{"text":{"type":"string","title":"Text"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"}},"type":"object","required":["text"],"title":"ComplaintCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Review":{"properties":{"text":{"type":"string","title":"Text"},"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"id":{"type":"integer","title":"Id"},"ad_id":{"type":"integer","title":"Ad Id"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["text","rating","id","ad_id","user_id","created_at"],"title":"Review"},"ReviewCreate":{"properties":{"text":{"type":"string","title":"Text"},"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"}},"type":"object","required":["text","rating"],"title":"ReviewCreate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"login":{"type":"string","title":"Login"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"id":{"type":"integer","title":"Id"},"is_admin":{"type":"boolean","title":"Is Admin"},"is_banned":{"type":"boolean","title":"Is Banned"},"password_hash":{"type":"string","title":"Password Hash"}},"type":"object","required":["login","id","is_admin","is_banned","password_hash"],"title":"User"},"UserBase":{"properties":{"login":{"type":"string","title":"Login"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","required":["login"],"title":"UserBase"},"UserCreate":{"properties":{"login":{"type":"string","title":"Login"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["login","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}